Add files via upload
***Duplicate Virus***
Here is a Simple batch virus that contains only 6 lines, and it has the
function to keep on creating a folder with the name “virus” until a user
stops it. It is an example of how this small code will consume the system’s

***dns poisoning malware***
This Batch file has the function of modifying the DNS Management of a
system by editing the hosts.txt file that resides inside
‘C:\windows\system32\drivers\ etc\hosts.txt’, so that it will take you to
some malicious websites instead of landing you on the legitimate
website. This may also be used for phishing, i.e. redirecting you to a
malicious website that looks exactly like the legitimate one, and then
stealing credentials.
resource.

***Application Bomber***
Application bomber is a superset of window bomber, this has a close
relation to the above-given fork bomber program, where this ‘application
bomber’ we don’t call the program using the name itself (simply known as
fork), whereas in this program we are going to open up several applications
continuously using a loop.

***Fork bombing:***
Most of them have heard about the word ‘fork()’, which is used to create
child process, likewise fork bombing, is nothing but calling a program by
itself again and again with an infinite loop making the system crash by
popping up hundreds of windows on the screen.

***msg annoyer***
Message Annoyer is a batch program that uses the same concept as
above but will interact with the user and anyhow annoying
them by popping up some message box containing the given messages
in it.

***Service disabler***
The following piece of code is used for stopping some critical windows
services.